+++ metadata.rb
@@ -4,6 +4,10 @@ maintainer_email "lloyd@firebellydesign.com"
 license          "GNU Public License 3.0"
 description      "Installs/Configures nginx_conf"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.0.7"
+version          "0.0.8"
+
 depends 'nginx'
-supports 'ubuntu'
+
+%w{ ubuntu debian centos redhat amazon scientific oracle fedora }.each do |os|
+  supports os
+end