+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 ## v0.0.7:
 
 * Use nginx binary attribute for nginx location.
+
+## v0.0.8:
+
+* removed available_sites_repo since we are integrating more with nginx cookbook
+* removed enabled_sites_repo since we are integrating more with nginx cookbook
+* added nginx_conf_options function to handle template configuration